Fixes #388 Vanilla iron blocks appearing during worldgen due to type cast.

+++ TFC_Shared/src/TFC/WorldGen/MapGenCaves256TFC.java
@@ -190,7 +190,7 @@ public class MapGenCaves256TFC extends MapGenBaseTFC
 											{
 												if (var50 < 10)
 												{
-													blockArray[index] = (byte)Block.lavaStill.blockID;
+													blockArray[index] = (short)Block.lavaStill.blockID;
 													metaArray[index] = 0;
 												}
 												else
